+++ ATTACK_BOT.hpp
@@ -92,7 +92,7 @@ public:
     // Clear all the dead units in the vector of attack_units
     void all_alive(std::vector<const Unit*> attack_units);
 
-    std::vector<Spotted_Enemy> last_seen_near(Point2D location, int radius, int since);
+    std::vector<Spotted_Enemy> last_seen_near(Point2D location, float radius, int since);
 
 private:
     std::vector<TF_unit> units;
