+++ DEFENCE_BOT.hpp
@@ -90,7 +90,7 @@ public:
 
     void setAgents(TF_Agent* attackb, TF_Agent* resourceb, TF_Agent* scoutb);
 
-    std::vector<Spotted_Enemy> last_seen_near(sc2::Point2D location, int radius, int since);
+    std::vector<Spotted_Enemy> last_seen_near(sc2::Point2D location, float radius, int since);
 
     void init();
