+++ RESOURCE_BOT.cpp
@@ -389,6 +389,6 @@ bool RESOURCE_BOT::buildCheckDuplicate(Units scvs, ABILITY_ID ability_to_build_s
     return false;
 }
 
-std::vector<Spotted_Enemy> RESOURCE_BOT::last_seen_near(Point2D location, int radius, int since) {
+std::vector<Spotted_Enemy> RESOURCE_BOT::last_seen_near(Point2D location, float radius, int since) {
     return scout->last_seen_near(location, radius, since);
 }
